Added circular refs and weak refs to callbacks
* Added weak reference support to callbacks.

Once the client drops all references to a callback, and the callback is
GC'd, the server will be notified of this and subsequently remove that
callback's id from the remote callback pool. This prevents the remote
callback pool from growing forever.

* Added circular reference support.

All objects sent as arguments to callbacks can now contain circular
references.

* Updated and improved documentation.